all - update qpp to main.
Three tests are failing.

Two of them because resetting an engine checks that the state
matches the circuit size, which fails if measurements happened.

The other one is commented as a bug from qpp. TODO: investigate.

More generally, qpp-fork should be synced with qpp, checking whether
the added commits are still relevant.